🔥🔥 actor: event rout + logger treatment
