+++ mini-app/src/server/db/events.ts
@@ -394,3 +394,37 @@ export const getEventById = async (eventId: number) => {
 
   return event === undefined || event.length === 0 ? null : event[0];
 };
+
+export const getEventsForSpecialRole = async (userRole: string, userId?: number) => {
+  if (userRole === "admin") {
+    return await db
+      .select()
+      .from(events)
+      .where(eq(events.hidden, false))
+      .orderBy(desc(events.created_at))
+      .execute();
+  } else if (userRole === "organizer" && userId) {
+    return await db
+      .select()
+      .from(events)
+      .where(and(eq(events.hidden, false), eq(events.owner, userId)))
+      .orderBy(desc(events.created_at))
+      .execute();
+  } else {
+    return [];
+  }
+};
+
+const eventDB = {
+  checkIsEventOwner,
+  checkIsAdminOrOrganizer,
+  checkEventTicketToCheckIn,
+  selectEventByUuid,
+  getUserEvents,
+  getOrganizerEvents,
+  getEventsWithFilters,
+  getEventByUuid,
+  getEventById,
+  getEventsForSpecialRole,
+};
+export default eventDB;
